fix: resolve all compilation errors after kPa→bar refactor and merge artifacts
- ckp.cpp: remove duplicate static var definitions (merge artifact)
- etb_driver.h/.cpp: rewrite without stm32h5xx.h, use project HAL (regs.h, timer.h, adc.h)
- torque_manager.h/.cpp: restore full content + add TorqueOutput/torque_manager_update
- etb_control.h/.cpp: restore full content + add EtbControlState/etb_control_update
- xtau_autocalib.cpp: replace ems::hal::millis() with millis() + add hal/system.h
- main_stm32.cpp: remove legacy float ETB block, fix xtau function names
- can_stack.cpp: fix status_bits type uint8_t→uint16_t to match header
- Makefile: add missing engine source files (map_estimator, xtau_autocalib, etb_control, torque_manager)

+++ src/app/can_stack.cpp
@@ -129,12 +129,12 @@ void can_stack_set_wbo2_rx_id(uint16_t id) noexcept {
 void can_stack_process(uint32_t now_ms,
                        const ems::drv::CkpSnapshot& ckp,
                        const ems::drv::SensorData&  sensors,
-                       int8_t  advance_deg,
-                       uint8_t pw_ms_x10,
-                       int8_t  stft_pct,
-                       uint8_t vvt_intake_pct,
-                       uint8_t vvt_exhaust_pct,
-                       uint8_t status_bits) noexcept {
+                       int8_t   advance_deg,
+                       uint8_t  pw_ms_x10,
+                       int8_t   stft_pct,
+                       uint8_t  vvt_intake_pct,
+                       uint8_t  vvt_exhaust_pct,
+                       uint16_t status_bits) noexcept {
     process_rx(now_ms);
 
     // Atualiza flag de falha após processar RX
